backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
celery[redis]==5.4.0
redis==5.2.0
docker==7.1.0
openai==1.55.0
web3==7.6.0
pinatapy-vourhey==0.2.0
aiosqlite==0.22.1
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
